refactor: extract DraggableThumbnail into widgets
