+++ lib/features/call/widgets/widgets.dart
@@ -4,3 +4,4 @@ export 'local_camera_preview_overlay.dart';
 export 'popup_menu.dart';
 export 'remote_video_view_overlay.dart';
 export 'rtc_stream_view.dart';
+export 'draggable_thumbnail.dart';
